LayoutTests/fast/forms/date/date-editable-components/date-editable-components-keyboard-events-expected.txt

Test for keyboard operations for <input type=date>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Digit keys
PASS input.value is "2012-09-02"
PASS changeEventsFired is 4
PASS inputEventsFired is 4

Digit keys with leading zero
PASS input.value is "0034-05-06"
PASS changeEventsFired is 3
PASS inputEventsFired is 3

Digit keys and backspace key
PASS input.value is "2020-02-04"
PASS changeEventsFired is 6
PASS inputEventsFired is 6

Digit keys with timeout
PASS input.value is "0001-02-04"
PASS changeEventsFired is 2
PASS inputEventsFired is 2

Digit keys clamp value
PASS input.value is "9999-12-31"
PASS changeEventsFired is 4
PASS inputEventsFired is 4

Left/Right arrow keys
PASS input.value is "0002-02-02"
PASS input.value is "0002-03-03"
PASS changeEventsFired is 3
PASS inputEventsFired is 3

Advance field keys
PASS input.value is "2020-06-02"
PASS input.value is "2020-06-03"
PASS input.value is "2020-06-04"
PASS input.value is "2020-06-05"
PASS input.value is "2020-06-06"
PASS input.value is "2020-06-07"
PASS input.value is "2020-06-08"
PASS changeEventsFired is 7
PASS inputEventsFired is 7

Up/Down arrow keys
PASS input.value is "2020-01-20"
PASS input.value is "2020-02-20"
PASS input.value is "2020-01-20"
PASS input.value is "2020-12-20"
PASS changeEventsFired is 4
PASS inputEventsFired is 4

Tab key
PASS input.value is "0002-02-02"
PASS document.activeElement.id is "after"
PASS input.value is "0002-03-03"
PASS document.activeElement.id is "before"
PASS changeEventsFired is 3
PASS inputEventsFired is 3

Backspace key
PASS input.value is ""
PASS input.value is "2020-07-26"
PASS changeEventsFired is 2
PASS inputEventsFired is 2

Delete key
PASS input.value is ""
PASS changeEventsFired is 1
PASS inputEventsFired is 1

Disabled/readonly
PASS input.value is "2020-09-01"
PASS input.value is "2020-01-01"
PASS input.value is "2020-01-01"
PASS input.value is "2020-01-02"
PASS changeEventsFired is 2
PASS inputEventsFired is 2
PASS successfullyParsed is true

TEST COMPLETE
  
